Apparently added some new materials? Bobbing work
I don't know why but unity says I have some materials added for our
models, so I'll push them in case they're important.  I also implemented
code for a bobbing boat, but it caused a conflict so I saved it to
desktop and we'll use it later if we want to.