Finally fixed controller support
Hey everybody, I fixed the controller support. Unfortunatey I had to buy
an editor extension to do it, which means I'm not allowed to share the
source code since that would technically be distributing illegal copies
of the editor extension I bought  :/
BUT, it's still fixed, so I'm uploading the built version of the game
for you guys to play and use on your portfolios and what not, so here it
is.